+++ service/register/model/user.go
@@ -5,8 +5,8 @@ type User_info struct {
 	ID               int64  `gorm:"column:id;primary_key;AUTO_INCREMENT;not null"`
 	Username         string `json:"username"`
 	Password_hash    string `json:"password_Hash"`
-	Follow_count     int
-	Follower_count   int
+	Follow_count     int64
+	Follower_count   int64
 	Favorates        []*Video //点赞的视频
 	Videos_published []*Video //发布的视频
 	//Comments         string   //发布的评论
